third_party/WebKit/LayoutTests/dom/domparsing/xmlserializer-doctype2-expected.txt

This tests XMLSerializer.serializeToString() on a newly created DocumentType node does not throw since the node has an associated document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS text = serializer.serializeToString(docType) did not throw exception.
PASS text is "<!DOCTYPE aDocTypeName PUBLIC \"aPublicID\" \"aSystemID\">"
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
